+++ backend/src/database/migrations/20221028134900_add_contact_project_constraint.js
@@ -1,0 +1,13 @@
+exports.up = function (knex) {
+  return knex.raw(`
+    ALTER TABLE data.contact_project
+    ADD CONSTRAINT project_contact_role_unique UNIQUE (project_id, contact_id, contact_role);
+  ;`);
+};
+
+exports.down = function (knex) {
+  return knex.raw(`
+    ALTER TABLE data.contact_project
+    DROP CONSTRAINT project_contact_role_unique;
+  ;`);
+};